Rename some task files
