+++ tasks/components/vendor_fonts.js
