+++ tasks/components/vendor_scripts.js
