+++ tasks/components/vendor_stylesheets.js
