+++ tasks/watch.js
@@ -1,11 +1,11 @@
-var clientStylesheets = require('./build/client_stylesheets');
-var clientTemplates = require('./build/client_templates');
-var clientScripts = require('./build/client_scripts');
-var vendorStylesheets = require('./build/vendor_stylesheets');
-var vendorScripts = require('./build/vendor_scripts');
-var vendorFonts = require('./build/vendor_fonts');
-var serverViews = require('./build/server_views');
-var serverApp = require('./build/server_app');
+var clientStylesheets = require('./components/client_stylesheets');
+var clientTemplates = require('./components/client_templates');
+var clientScripts = require('./components/client_scripts');
+var vendorStylesheets = require('./components/vendor_stylesheets');
+var vendorScripts = require('./components/vendor_scripts');
+var vendorFonts = require('./components/vendor_fonts');
+var serverViews = require('./components/server_views');
+var serverScripts = require('./components/server_scripts');
 
 var Q = require('q');
